requirements.txt
Django
djangorestframework>=3.14.0,<4.0
django-allauth
django-cors-headers
django-crispy-forms
